Generate a basic District Ball image

Generate a basic District Ball image by specifying parameters for the district and the expression of the ball.

generate_ball(
  district = "blank",
  expression = "happy",
  scale = 200,
  vjust = 1,
  hjust = 1,
  bg_colour = "#FFFFFF",
  caption = "",
  caption_colour = "#000000",
  caption_size = 20
)

Arguments

district
character string specifying the district. Run district_options() to find out all the valid options for this argument.

expression
character string specifying the facial expression. Run expression_options() to find out all the valid options for this argument.

scale
numeric value. Size of the image to scale to. Defaults to 200.

vjust
numeric value. Adjust the vertical position of the expression.

hjust
numeric value. Adjust the horizontal of the expression.

bg_colour
character string. HEX code for the background colour.

caption
character string. Caption to appear at top of image. Defaults to blank.

caption_colour
character string. HEX code for the caption colour.

caption_size
numeric value. Text size of the caption.

Examples

if (FALSE) {
# Run a Kowloon City ball, fierce expression, scale to 500
generate_ball("kowloon-city",
              "fierce",
              scale = 500,
              vjust = 1.0,
              caption = "We need cleaner streets!",
              caption_size = 30)
}